% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc_mode.r
\name{fun_calc_mode}
\alias{fun_calc_mode}
\title{Calculates the mode of the provided data vector}
\usage{
fun_calc_mode(x)
}
\arguments{
\item{x}{data vector.}
}
\value{
mode value.
}
\description{
Calculates the mode (most frequent value) of the provided data
vector.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
