% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_center_image.r
\name{fun_center_image}
\alias{fun_center_image}
\title{Centers the provided image}
\usage{
fun_center_image(image, hdrlst, header, sdo.image = "FALSE")
}
\arguments{
\item{image}{tibble image with i and j pixel coordinates and pixel values x.
Additional columns are ignored and will be not returned.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{sdo.image}{boolean switch for dummy use in the case of non sdo calcium
images.}
}
\value{
tibble with centered image.
}
\description{
Centers the provided image by shifting the x-axis and the
y-axis of the image appropriately. Do not center sdo images, since they
are centered already. The algorithm is slightly adapted from
Berry, R. and Burnell, J.: The Handbook of Astronomical Image Processing,
2nd edition 2005, p. 323.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
