% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_clv_correct.r
\name{fun_clv_correct}
\alias{fun_clv_correct}
\title{Flattens an image}
\usage{
fun_clv_correct(x, name.image = "calib", name.clv = "sclv")
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j and columns
with the pixel values of the image and the clv function.}

\item{name.image}{name of the column containing the image pixel values.}

\item{name.clv}{name of the column containing the clv function values.}
}
\value{
tibble with additional column to x containing the flattened image.
}
\description{
Flattens a solar image according to a provided clv function.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
