% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_clv_function.r
\name{fun_clv_function}
\alias{fun_clv_function}
\title{Calculates the standardized center-to-limb-variation function}
\usage{
fun_clv_function(
  x,
  clv.i0,
  sclv.method = "NL",
  sdo.image = "FALSE",
  clv.coeff = NULL
)
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j, pixel
values x, pixel coordinates xi, yj and angle theta from an image disc.}

\item{clv.i0}{central intensity of the fitted clv function of the image.}

\item{sclv.method}{implemented are "Allen", "PS", "NL" and "fit".}

\item{sdo.image}{boolean switch for dummy use in the case of non sdo calcium
images.}

\item{clv.coeff}{vector with 6 coefficients from the fitted clv function.}
}
\value{
z tibble with additional column to x containing the standardized clv
function and the cf vector with coefficients of the standardized clv
function.
}
\description{
Calculates the center-to-limb-variation function according to
formulae published by different authors, including Allen in Cox (2000),
Pierce & Slaughter (1977) and Neckel & Labs (1994). Furthermore, a customly
fitted clv may be calculated. Note, that this is the only available option
for SDO images, since near the temperature minimum, a near flat clv is
expected but the SDO images show a clear limb darkening.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
