% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_clv_rmplane.r
\name{fun_clv_rmplane}
\alias{fun_clv_rmplane}
\title{Subtracts plane from image}
\usage{
fun_clv_rmplane(x, hdrlst, coeff)
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j, pixel
values x from the original image, fill from the disc mask, image from the 
extracted image disc and theta.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{coeff}{containing the previously fitted coefficients of an 
eight-parameter fit function.}
}
\value{
tibble containing the corrected image values.
}
\description{
Subtracts a previously fitted plane from the provided image
  values.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
