% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_correct_gamma.r
\name{fun_correct_gamma}
\alias{fun_correct_gamma}
\title{Corrects gamma of an image}
\usage{
fun_correct_gamma(image, hdrlst, header, gamma = 1.5, correct.gamma = "FALSE")
}
\arguments{
\item{image}{tibble image with i and j pixel coordinates and pixel values x.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{gamma}{num gamma correction factor. For flat SDO images a gamma of 1.5
is recommended.}

\item{correct.gamma}{boolean if TRUE the image will be gamma corrected.}
}
\value{
list with image tibble, parsed header and full header vectors.
}
\description{
Corrects gamma and returns an image with the provided
bit depth.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
