% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cut_image.r
\name{fun_cut_image}
\alias{fun_cut_image}
\title{Extracts central part of image without text in the corners}
\usage{
fun_cut_image(
  x,
  hdrlst,
  header,
  cut.image,
  cut.threshold = 20,
  cut.method = "relative",
  cut.border.pix = 100
)
}
\arguments{
\item{x}{tibble containing 3 columns with pixel coordinates i and j and pixel
intensity values x.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{cut.image}{image contains text in the corners.}

\item{cut.threshold}{cutoff threshold for images with text in the corners
in absolute or relative units.}

\item{cut.method}{If cut.method = 'absolute' then the cutoff threshold value
for images with text in the corners is in ADU units. Otherwise it is the
cut.threshold / 100 fraction of the maximal ADU value given by 2^bitpix.}

\item{cut.border.pix}{number of pixels to be added to the radius of the
circle to extract the image without the text in the corners.}
}
\value{
list with disc.image, hdrlst and header.
}
\description{
If the original image contains additional text in the image
corners, the approximate center coordinates and the approximate radius of
the image disc are estimated and the image disk is extracted with an
oversized mask. The size of the original image and the number of bits per
pixel are taken from the FITS header provided by hdrlst.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
