% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_date2jd.r
\name{fun_date2jd}
\alias{fun_date2jd}
\title{Calculates Julian Date from a vector containing date with time objects}
\usage{
fun_date2jd(date_time)
}
\arguments{
\item{date_time}{vector containing date and time in FITS format.}
}
\value{
vector with Julian Dates.
}
\description{
Calculates Julian Date from a vector containing date with time
objects. Does not work for negative years. Requires 4 digits for year.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
