% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_date_grid.r
\name{fun_date_grid}
\alias{fun_date_grid}
\title{Provides a tibble with dates from start to end}
\usage{
fun_date_grid(start_date, end_date, granularity = "day")
}
\arguments{
\item{start_date}{First date of grid.}

\item{end_date}{Last date of grid.}

\item{granularity}{Implemented are "year", "month" and "day".}
}
\value{
tibble containing date, and depending on the granularity also year,
month and day.
}
\description{
Provides a tibble with dates and columns containing year, month
and day. The granularity can be chosen from year month or day. Works for
all years within the Gregorian calendar.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
