% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_deltaT.r
\name{fun_deltaT}
\alias{fun_deltaT}
\title{Calculates delta T in seconds according to formulae by Fred Espenak}
\usage{
fun_deltaT(date_time)
}
\arguments{
\item{date_time}{vector containing date and time in FITS format.}
}
\value{
value of delta T in seconds.
}
\description{
Calculates delta T = TD - UT in seconds according to formulae
given by Fred Espenak. Before 2005, the polynomial expressions for delta T
are from the Five Millennium Canon of Solar Eclipses: -1999 to +3000.
Source: http://www.eclipsewise.com/help/deltatpoly2014.html
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
