% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_disc_center.r
\name{fun_disc_center}
\alias{fun_disc_center}
\title{Calculates disc center coordinates and radius of the indicator disc}
\usage{
fun_disc_center(x, disc = "fill")
}
\arguments{
\item{x}{tibble containing at least 3 columns with pixel coordinates i and j
and disc indicator values disk.}

\item{disc}{name of variable in x with disc indicator values.}
}
\value{
tibble with intensity center coordinates x_i for image columns
and y_j for image rows and radius of the disc indicator circle.
}
\description{
Calculates disc center coordinates and the radius for a given
disc indicator.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
