% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_disc_quality.r
\name{fun_disc_quality}
\alias{fun_disc_quality}
\title{Calculates the image quality as standard error of the border pixel}
\usage{
fun_disc_quality(x, hdrlst, disc.center, border = "border")
}
\arguments{
\item{x}{tibble containing at least 3 columns with pixel coordinates i and j
and border indicator values.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{disc.center}{tibble with coordinates x_i for image columns and
y_j for image rows of the disc center and radius r of the disc.}

\item{border}{name of variable in x with border indicator values.}
}
\value{
tibble data containing for each border pixel the calculated radius
and the residuals against the provided disc radius. Furthermore a
tibble sigma.border including the mean radius, the standard deviation
of the mean, the standard deviation of a single measurement in pixels and
in arcseconds.
}
\description{
Calculates the image quality as standard deviation of the
individual disc radius determinations of each border pixel expressed both
in pixels and in arcseconds.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
