% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_disc_rings.r
\name{fun_disc_rings}
\alias{fun_disc_rings}
\title{Assigns ring number to disc pixels}
\usage{
fun_disc_rings(x, num.rings = 50)
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j and pixel
values x from the original image.}

\item{num.rings}{number of rings counted from limb to disc center.}
}
\value{
z tibble with additional column to x containing the ring number.
}
\description{
Computes a table with the ring borders given the number of rings
and assigns the number of the ring for each disc pixel.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
