% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_exchange_header.r
\name{fun_exchange_header}
\alias{fun_exchange_header}
\title{Exchanges the header of a single FITS file}
\usage{
fun_exchange_header(
  im,
  exchange.header = "FALSE",
  inp_hdata_path,
  inp_hfile_name
)
}
\arguments{
\item{im}{list with image matrix, parsed header and full header vectors.}

\item{exchange.header}{boolean if TRUE the header of the original image will
be changed.}

\item{inp_hdata_path}{string with input data path for new header file.}

\item{inp_hfile_name}{string with input file name of the new header file.}
}
\value{
list with image matrix, exchanged parsed header and exchanged full
header vectors.
}
\description{
Exchanges existing image header of a single FITS file with the
header from another single FITS file.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
