% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_extract_row.r
\name{fun_extract_row}
\alias{fun_extract_row}
\title{Extracts a single row of image frame}
\usage{
fun_extract_row(x, row = NULL)
}
\arguments{
\item{x}{tibble containing 3 columns wih pixel coordinates i and j and pixel
intensity values x.}

\item{row}{int index of row to be extracted.}
}
\value{
tibble containing 2 columns wih column numbers i and pixel
  intensity values x.
}
\description{
Extracts a single row of image frame.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
