% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_flip_image.r
\name{fun_flip_image}
\alias{fun_flip_image}
\title{Flips the provided image}
\usage{
fun_flip_image(image, hdrlst, header, sdo.image = "FALSE")
}
\arguments{
\item{image}{tibble image with i and j pixel coordinates and pixel values x.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{sdo.image}{boolean switch for dummy use in the case of non sdo calcium
images.}
}
\value{
tibble with flipped image, updated hdrlst and header.
}
\description{
Flips the provided image by reversing the y-axis of the image 
  (interchanging up and down). Do not flip for sdo images. Algorithm from
  Berry, R. and Burnell, J.: The Handbook of Astronomical Image Processing, 
  2nd edition 2005, p. 331f.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
