% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_grid_plt.r
\name{fun_grid_plt}
\alias{fun_grid_plt}
\title{Adds grid to an image}
\usage{
fun_grid_plt(x, hdrlst)
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j, pixel
values x from the image and a binarised mask with the matching grid.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{theta}{angle measured in degrees from the disc center to the limb.}

\item{num.rings}{number of rings counted from limb to disc center.}
}
\value{
tibble with additional column containing the image with ring borders.
}
\description{
Adds grid to an image, both provided by the same tibble.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
