% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_jd2date.r
\name{fun_jd2date}
\alias{fun_jd2date}
\title{Calculates date and time vector from Julian Dates vector}
\usage{
fun_jd2date(JD)
}
\value{
vector containing date with time in FITS format.
}
\description{
Calculates date and time vector from vector containing Julian
  Dates objects. Does not work for negative years.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
