% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_filename.r
\name{fun_make_filename}
\alias{fun_make_filename}
\title{Constructs a full image file name as date_ut_file_name_ext}
\usage{
fun_make_filename(date, ut, file_name_ext)
}
\arguments{
\item{date}{Character date vector containing YYYY-MM-DD.}

\item{ut}{Character time vector containing HHMM.}

\item{file_name_ext}{Character vector containing file name particel and file
extension.}
}
\value{
character vector containing resulting image filename.
}
\description{
Constructs a full image file name as date_ut_file_name_ext.
The resulting file name is optimised for the download from jsoc.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
