% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_path.r
\name{fun_make_path}
\alias{fun_make_path}
\title{Constructs a full image path with root_path part and daily subfolder
structure. The image file name is constructed as date_ut_file_name_ext}
\usage{
fun_make_path(date, ut, root_path, file_name_ext)
}
\arguments{
\item{date}{Character date vector containing YYYY-MM-DD.}

\item{ut}{Character time vector containing HHMM.}

\item{root_path}{Character vector containing root path.}

\item{file_name_ext}{Character vector containing file name particel and file
extension.}
}
\value{
character vector containing resulting full image path.
}
\description{
Constructs a full image path with root_path part and daily
subfolder structure. The image file name is constructed as
date_ut_file_name_ext. The resulting path is optimised for the download of
image files from jsoc.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
