% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_root_path_date.r
\name{fun_make_root_path_date}
\alias{fun_make_root_path_date}
\title{Constructs an extended image root path with root_path part and
daily subfolder structure.}
\usage{
fun_make_root_path_date(date, root_path)
}
\arguments{
\item{date}{Character date vector containing YYYY-MM-DD.}

\item{root_path}{Character vector containing root path.}
}
\value{
character vector containing resulting extended root path.
}
\description{
Constructs an extended image root path with root_path part and
daily subfolder structure. The resulting path is optimised for the download
of image files from jsoc.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
