% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mask_border.r
\name{fun_mask_border}
\alias{fun_mask_border}
\title{Marks the border pixels of a given mask}
\usage{
fun_mask_border(x, mask = "th")
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j and mask.}

\item{mask}{name of the column in x containing the mask.}
}
\value{
z tibble with additional column to x containing the border pixels.
}
\description{
Marks the border pixels of a given mask.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
