% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mask_circle.r
\name{fun_mask_circle}
\alias{fun_mask_circle}
\title{Determines a circle given the center coordinates and the radius}
\usage{
fun_mask_circle(x, disc.center, border.pix = 0)
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j.}

\item{disc.center}{tibble with center coordinates x_i for image columns and
y_j for image rows and radius of the circle.}

\item{border.pix}{number of pixels to be added to the radius of the circle.}
}
\value{
z tibble with additional column to x containing the limb pixels of
the circle.
}
\description{
Determines limb pixels of a circle with given center
coordinates and radius and stores them in a mask. Optionally,
additional border pixels may be added to the circle radius.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
