% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mask_diff.r
\name{fun_mask_diff}
\alias{fun_mask_diff}
\title{Marks all pixels between two given thresholds}
\usage{
fun_mask_diff(
  x,
  hdrlst,
  lower.threshold = 5,
  upper.threshold = 10,
  method = "relative"
)
}
\arguments{
\item{x}{tibble containing 3 columns wih pixel coordinates i and j and pixel
intensity values x.}

\item{hdrlst}{tibble containing image FITS header keywords and values.}

\item{lower.threshold}{cutoff threshold in absolute or relative units.}

\item{upper.threshold}{cutoff threshold in absolute or relative units.}

\item{method}{If method = 'absolute' then the cutoff thereshold value is in
ADU units. Otherwise it is the threshold / 100 fraction of the maximal ADU
value given by 2^bitpix.}
}
\value{
list containing tibble with 4 columns for i, j pixel coordinates, 
  intensity value x, binarised mask th and tibbles with mask.lower.threshold
  and mask.upper.threshold values.
}
\description{
Calculates binarised pixel mask for all pixels above or equal
  a given lower.threshold but lower than a given upper.threshold.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
