% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_match_sdo_image.r
\name{fun_match_sdo_image}
\alias{fun_match_sdo_image}
\title{Searches for the best matching sdo image of a provided observing time}
\usage{
fun_match_sdo_image(url, time)
}
\arguments{
\item{url}{Character vector containing root path to observing day.}

\item{time}{Character time vector containing HHMM.}
}
\value{
character vector containing resulting path to matching SDO image.
}
\description{
Searches for the best matching sdo image of a provided observing
time. The resulting path is optimised for the download of image files from
jsoc. Some of the code was provided by chatGPT.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
