% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_minimal_header.r
\name{fun_minimal_header}
\alias{fun_minimal_header}
\title{Constructs minimal image header}
\usage{
fun_minimal_header(x, bitpix = 16, header = "")
}
\arguments{
\item{x}{image data matrix.}

\item{bitpix}{equals to 16 or 8 corresponding to 16 bit or 8 bit images.}

\item{header}{string vector with card images of an existing header.}
}
\value{
character vector with FITS header keywords.
}
\description{
Constructs character vector with minimal FITS header keywords.
Adds keywords as provided by card images of an existing header.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
