% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_parse_file_ext.r
\name{fun_parse_file_ext}
\alias{fun_parse_file_ext}
\title{Parses file extension}
\usage{
fun_parse_file_ext(file.name)
}
\arguments{
\item{file.name}{string original image file name.}
}
\value{
string parsed file extension.
}
\description{
Parses file extension from the provided file name.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
