% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_parse_filename.r
\name{fun_parse_filename}
\alias{fun_parse_filename}
\title{Parses file names for mandatory keyword DATE-OBS}
\usage{
fun_parse_filename(file.name, parse.method = "SDO/HMI")
}
\arguments{
\item{file.name}{string original image file name.}

\item{parse.method}{string instrument specific parsing method. Implemented
are the SDO/HMI naming convention of Intensity based files from JSOC.}
}
\value{
list with parsed date.obs.
}
\description{
Reconstructs mandatory FITS keywords from the provided file
name.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
