% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_parse_header.r
\name{fun_parse_header}
\alias{fun_parse_header}
\title{Parses FITS header}
\usage{
fun_parse_header(headerName)
}
\arguments{
\item{headerName}{list containing image FITS header.}
}
\value{
Character vector with parsed FITS header entries.
}
\description{
Parses FITS header. COMMENT, HISTORY, empty and some preserved
keywords are omitted.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
