% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_parse_keywords.r
\name{fun_parse_keywords}
\alias{fun_parse_keywords}
\title{Updates the header and hdrlst with date and time information}
\usage{
fun_parse_keywords(
  file.name,
  header,
  hdrlst,
  parse.filename = "FALSE",
  parse.method = "SDO/HMI"
)
}
\arguments{
\item{file.name}{string original image file name.}

\item{header}{list containing image FITS header.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{parse.filename}{boolean switch for dummy use for frames with
existing fits keywords.}

\item{parse.method}{string method for parsing the provided file.name.}
}
\value{
list with updated header and hdrlst.
}
\description{
Reconstructs the FITS keyword DATE-OBS from the provided file
name. Updates the header and hdrlst.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
