% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_parse_obs_time.r
\name{fun_parse_obs_time}
\alias{fun_parse_obs_time}
\title{Extracts observation time from jsoc-type file names}
\usage{
fun_parse_obs_time(file.name)
}
\arguments{
\item{file.name}{character string with full file name.}
}
\value{
character string with extracted observation time.
}
\description{
Extracts observation time from jsoc-type file names.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
