% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_header.r
\name{fun_read_header}
\alias{fun_read_header}
\title{Reads FITS ASCII header from disc}
\usage{
fun_read_header(zz, maxLines = 5000)
}
\arguments{
\item{zz}{binary port to file.}

\item{maxLines}{maximal number of header lines to search for END statement.
May be increased for very large headers.}
}
\value{
Character vector with FITS header card images.
}
\description{
Reads FITS ASCII header from disc.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
