% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_read_image.r
\name{fun_read_image}
\alias{fun_read_image}
\title{Reads FITS image from disc}
\usage{
fun_read_image(filename = "sunxplrr.fits", maxLines = 5000)
}
\arguments{
\item{filename}{output path and file name.}

\item{maxLines}{maximal number of header lines to search for END statement.
May be increased for very large headers.}
}
\value{
list with image matrix, parsed header and full header vectors.
}
\description{
Reads 16 bit or 8 bit FITS image from disc.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
