% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_readFITSarray.r
\name{fun_readFITSarray}
\alias{fun_readFITSarray}
\title{Reads FITS binary array from disc}
\usage{
fun_readFITSarray(zz, hdr)
}
\arguments{
\item{zz}{binary port to file.}

\item{hdr}{parsed header keyword and values vector.}
}
\value{
list with image matrix and parsed header vector.
}
\description{
Reads 16 bit or 8 bit FITS binary array from disc.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
