% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ring_get.r
\name{fun_ring_get}
\alias{fun_ring_get}
\title{Returns ring parameters given a theta value}
\usage{
fun_ring_get(theta = 75, num.rings = 50)
}
\arguments{
\item{theta}{angle measured in degrees from the disc center to the limb.}

\item{num.rings}{number of rings counted from limb to disc center.}
}
\value{
table and a vector with ring characteristic for a given theta value.
}
\description{
Computes a table with the ring borders given the number of rings
and returns the ring characteristics given a single theta value.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
