% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ring_plt.r
\name{fun_ring_plt}
\alias{fun_ring_plt}
\title{Adds a ring to the image given a theta value}
\usage{
fun_ring_plt(x, hdrlst, theta = 75, num.rings = 50)
}
\arguments{
\item{x}{tibble containing columns with pixel coordinates i and j, pixel
values x from the original image, fill from the disc mask, image from the
extracted image disc and theta.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{theta}{angle measured in degrees from the disc center to the limb.}

\item{num.rings}{number of rings counted from limb to disc center.}
}
\value{
tibble with additional column containing the image with ring borders.
}
\description{
Computes the ring characteristics given a single theta value and
adds the inner and outer ring borders to the image.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
