% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_tibbl2mat.r
\name{fun_tibbl2mat}
\alias{fun_tibbl2mat}
\title{Umwandlung einer tibble oder eines data.frames in eine Matrix}
\usage{
fun_tibbl2mat(x)
}
\arguments{
\item{x}{tibble with 3 columns for i, j pixel coordinates and intensity 
value x.}
}
\value{
matrix with i columns and j rows.
}
\description{
Wandelt ein tibble oder ein data.frame in eine Matrix um.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
