% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_tile_jpg_image.r
\name{fun_tile_jpg_image}
\alias{fun_tile_jpg_image}
\title{Reads JPG image from disc}
\usage{
fun_tile_jpg_image(filename = "sun_logo.jpg", out_data_path)
}
\arguments{
\item{filename}{input path and file name.}

\item{bitpix}{int image will be scaled to the provided bit depth.}
}
\value{
list with image matrix, parsed header and full header vectors.
}
\description{
Reads monochrome or three color JPG image from disc. Constructs
minimal FITS header and returns image with the provided bit depth.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
