% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_write_image.r
\name{fun_write_image}
\alias{fun_write_image}
\title{Writes FITS image to disc}
\usage{
fun_write_image(x, file, hdrlst, header)
}
\arguments{
\item{x}{image data matrix.}

\item{file}{output path and file name.}

\item{hdrlst}{tibble with header keywords and values.}

\item{header}{string vector with card images of the existing header and of
new or modified card images.}
}
\value{
nothing.
}
\description{
Writes 16 bit or 8 bit FITS image to disc.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
