% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clv_correction.r
\name{mod_clv_correction}
\alias{mod_clv_correction}
\title{Fits the clv function, calibrate it and flattens the image}
\usage{
mod_clv_correction(
  x,
  hdrlst,
  header,
  model = "poly_with_plane",
  run = 3,
  clip.resid.out = "FALSE",
  sclv.method = "NL",
  sdo.image = "FALSE",
  light.save = "FALSE"
)
}
\arguments{
\item{x}{tibble containing columns wih pixel coordinates i and j and all
the calculated image information as provided by the disc extraction module.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{model}{implemented are "poly" for polynomial fit with 6 coefficients,
"poly_with_plane" for polynomial fit with 6 coefficients and an additional
2 parameter plane through the disc center and a third model named as
"poly_with_plane_and_nonparametric_background" for polynomial fit with
6 coefficients, an additional 2 parameter plane and a nonparametric surface
fit in a gam structure.}

\item{run}{indicates how many iterations for spot and faculae elimination
should be run. No more than 3 iterations are implemented.}

\item{clip.resid.out}{if "TRUE" the clipped residuals from the last fit
iteration are returned. In batch mode this output is not needed.}

\item{sdo.image}{boolean switch for dummy use in the case of non sdo calcium
images.}

\item{light.save}{if TRUE only a small selection of csv files are saved.}
}
\value{
tibble containing clv function, calib image and flat image.
}
\description{
Fits a six parameter polynom to the limb-variation-function.
Additionnally, a plane may be fitted trough the center of the image disc.
Removes all fitted non clv variation from the image and flattens it to
intensity contrasts. Be aware, that the resulting intensity contrast values
should be multiplied with some appropriate factor before converting them
back to FITS, since in the FITS image all values are treated as integers.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
