% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_feature_extraction.r
\name{mod_feature_extraction}
\alias{mod_feature_extraction}
\title{Extracts the plage, enhanced network and active network features}
\usage{
mod_feature_extraction(
  x,
  hdrlst,
  header,
  plage.contrast = 1.35,
  en.contrast = 1.25,
  qn.contrast = 1.1
)
}
\arguments{
\item{x}{tibble containing 3 columns wih pixel coordinates i and j and flat 
image pixel values.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{plage.contrast}{cutoff threshold for plages.}

\item{en.contrast}{cutoff threshold for enhanced network.}

\item{qt.contrast}{cutoff threshold for quiet network.}
}
\value{
tibble with mask of the extracted features and their contrast values.
}
\description{
Extracts the plage, enhanced network and active network features 
  by applying cutoff thresholds. Calculates also the hemispheric corrected 
  values for the extracted feature areas and contrasts.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
