% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_param.r
\name{mod_load_param}
\alias{mod_load_param}
\title{Sets the parameter list for STU and SDO image analysis}
\usage{
mod_load_param(
  inp_file_name,
  sdo.image,
  inp_data_path,
  out_data_path,
  rds.output = "FALSE",
  full.output = "FALSE",
  light.save = "FALSE",
  fits.save = "FALSE",
  jpg.save = "FALSE"
)
}
\arguments{
\item{inp_file_name}{input file name with extension.}

\item{sdo.image}{if TRUE, param.lst for SDO image analysis is chosen.}

\item{inp_data_path}{full path to input directory.}

\item{out_data_path}{full path to output directory.}

\item{rds.output}{if TRUE all results are saved as R data file.}

\item{full.output}{if TRUE the full image data table is saved as csv file.}

\item{light.save}{if TRUE only a small selection of csv files are saved.}

\item{fits.save}{if TRUE some fits images are saved.}

\item{jpg.save}{if TRUE some daily charts are saved as jpg files.}
}
\value{
list containing param.lst.
}
\description{
Sets the parameter list for STU and SDO image analysis. The
add.border.pix and contrast thresholds are set following the recommendation
of Jannine Meier from 13.01.2020.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
