% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output.r
\name{mod_output}
\alias{mod_output}
\title{Saves results to disk}
\usage{
mod_output(
  images,
  hdrlst,
  header,
  total.indices,
  hemisphere.indices,
  latitude.indices,
  chart.indices,
  charts,
  synopsis.indices,
  synopsis,
  inp_file_name,
  rds.output = "FALSE",
  full.output = "FALSE",
  light.save = "FALSE",
  fits.save = "FALSE",
  jpg.save = "FALSE",
  out_data_path
)
}
\arguments{
\item{images}{tibble containing the previously calculated images.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{rds.output}{if TRUE all results are saved as R data file.}

\item{full.output}{if TRUE the full image data table is saved as csv file.}

\item{light.save}{if TRUE only a small selection of csv files are saved.}

\item{fits.save}{if TRUE some fits images are saved.}

\item{jpg.save}{if TRUE some daily charts are saved as jpg files.}

\item{out_data_path}{full path to output directory.}
}
\value{
list containing images, hdrlst, header, indices and charts.
}
\description{
Saves results to disk.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
