% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spot_calculation.r
\name{mod_spot_calculation}
\alias{mod_spot_calculation}
\title{Calculates activity indices and provides data for images and charts}
\usage{
mod_spot_calculation(x, hdrlst, header)
}
\arguments{
\item{x}{tibble containing the previously calculated images.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}
}
\value{
list with images, hdrlst, header, indices, charts and syopsis.
}
\description{
Calculates activity indices and provides input data for FITS
images and charts.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
