% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_modify_header.r
\name{newKwv}
\alias{newKwv}
\title{Utilities for modifying FITS header entries}
\usage{
newKwv(keyw, val, note = "")
}
\arguments{
\item{keyw}{FITS keyword name.}

\item{val}{FITS keyword value.}

\item{note}{FITS keyword description.}
}
\value{
string vector with modified header.
}
\description{
Utilities for modifying FITS header entries.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
