% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunxplrr-package.R
\docType{package}
\name{sunxplrr-package}
\alias{sunxplrr}
\alias{sunxplrr-package}
\title{sunxplrr: Tools for exploring Digital Images of the Sun}
\description{
Tools for solar image analysis, feature extraction, indices calculation and visualisation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wolfinstitute/sunxplrr}
  \item Report bugs at \url{https://github.com/wolfinstitute/sunxplrr/issues}
}

}
\author{
\strong{Maintainer}: Thomas K. Friedli \email{thomas.k.friedli@bluewin.ch} (\href{https://orcid.org/0000-0002-2086-7192}{ORCID}) [copyright holder]

}
\keyword{internal}
