% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_mod_calcium.r
\name{wrap_mod_calcium}
\alias{wrap_mod_calcium}
\title{Runs the sunviewr modules for the calcium image analysis}
\usage{
wrap_mod_calcium(
  inp_file_name,
  sdo.image,
  inp_data_path,
  out_data_path,
  rds.output = "FALSE",
  full.output = "FALSE",
  light.save = "FALSE",
  fits.save = "FALSE",
  jpg.save = "FALSE"
)
}
\arguments{
\item{inp_file_name}{input file name with extension.}

\item{sdo.image}{if TRUE, param.lst for SDO image analysis is chosen.}

\item{inp_data_path}{full path to input directory.}

\item{out_data_path}{full path to output directory.}

\item{rds.output}{if TRUE all results are saved as R data file.}

\item{full.output}{if TRUE the full image data table is saved as csv file.}

\item{light.save}{if TRUE only a small selection of csv files are saved.}

\item{fits.output}{if TRUE some fits images are saved.}

\item{jpg.output}{if TRUE some daily charts are saved as jpg files.}
}
\value{
list with resulting image and header.
}
\description{
Runs the sunviewr modules for the exploration of calcium images 
  and the determination of activity indices for one image.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
