% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_mod_jpg_sdo.r
\name{wrap_mod_jpg_sdo}
\alias{wrap_mod_jpg_sdo}
\title{Runs the sunxplrr modules for the SDO whitelight jpg image analysis}
\usage{
wrap_mod_jpg_sdo(
  inp_file_name,
  sdo.image,
  inp_data_path,
  out_data_path,
  rds.output = "FALSE",
  full.output = "FALSE",
  light.save = "FALSE",
  fits.save = "FALSE",
  jpg.save = "FALSE"
)
}
\arguments{
\item{inp_file_name}{input file name with extension.}

\item{sdo.image}{if TRUE, param.lst for SDO image analysis is chosen.}

\item{inp_data_path}{full path to input directory.}

\item{out_data_path}{full path to output directory.}

\item{rds.output}{if TRUE all results are saved as R data file.}

\item{full.output}{if TRUE the full image data table is saved as csv file.}

\item{light.save}{if TRUE only a small selection of csv files are saved.}

\item{fits.save}{if TRUE some fits images are saved.}

\item{jpg.save}{if TRUE some daily charts are saved as jpg files.}

\item{exchange.header}{boolean if TRUE the header of the original image will
be changed.}

\item{inp_hfile_name}{string with input file name of the new header file.}
}
\value{
list with resulting image and header.
}
\description{
Runs the sunxplrr modules for the exploration of whitelight
jpg images as provided by the jsoc server of SDO.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
