% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mask_create.r
\name{fun_mask_create}
\alias{fun_mask_create}
\title{Marks all pixels above or equal a given threshold}
\usage{
fun_mask_create(x, hdrlst, threshold = 10, method = "relative")
}
\arguments{
\item{x}{tibble containing 3 columns wih pixel coordinates i and j and pixel
intensity values x.}

\item{hdrlst}{tibble containing image FITS header keywords and values.}

\item{threshold}{cutoff threshold in absolute or relative units.}

\item{method}{If method = 'absolute' then the cutoff thereshold value is in
ADU units. Otherwise it is the threshold / 100 fraction of the maximal ADU
value given by 2^bitpix.}
}
\value{
list containing tibble with 4 columns for i, j pixel coordinates,
intensity value x and binary mask th and tibble with mask.threshold value.
}
\description{
Calculates binarised pixel mask according to a given thereshold.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
