% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sdo_keywords.r
\name{fun_sdo_keywords}
\alias{fun_sdo_keywords}
\title{Updates the header and hdrlst with sdo-specific keywords}
\usage{
fun_sdo_keywords(file.name, header, hdrlst, sdo.image = "FALSE")
}
\arguments{
\item{header}{list containing image FITS header.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{sdo.image}{boolean switch for dummy use in the case of non sdo calcium
images.}

\item{filename}{original file name of sdo image.}
}
\value{
list with updated header and hdrlst.
}
\description{
Reconstructs the FITS keywords DATE-OBS, NAXISn and WAVELNTH 
  from the provided sdo file name. Updates the header and hdrlst with some 
  other missing keywords.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
