% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_sun_ephem.r
\name{fun_sun_ephem}
\alias{fun_sun_ephem}
\title{Calculates ephemeris for physical coordinates of the Sun}
\usage{
fun_sun_ephem(hdrlst, sdo.image = "FALSE")
}
\arguments{
\item{hdrlst}{list containing image FITS header keywords and values.}

\item{sdo.image}{boolean switch for case of sdo image.}
}
\value{
tibble with P0, B0, L0, the Sun's apparent diameter in arcsecs and
  the Carrington rotation number.
}
\description{
Calculates ephemeris for physical coordinates of the Sun, 
  including P0, B0, L0, the Sun's apparent diameter in arcsecs and the 
  Carrington rotation number. 
  Formulae from: Meeus,J., Astronomical Algorithms, Willmann-Bell, 1991.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
