% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_tibbl2mat.r
\name{fun_tibbl2mat}
\alias{fun_tibbl2mat}
\title{Transforms a tibble or a data.frame to a matrix}
\usage{
fun_tibbl2mat(x)
}
\arguments{
\item{x}{tibble with 3 columns for i, j pixel coordinates and intensity
value x.}
}
\value{
matrix with i columns and j rows.
}
\description{
Transforms a tibble or a data.frame to a matrix.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
