% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_disc_flat.r
\name{mod_disc_flat}
\alias{mod_disc_flat}
\title{Recalibrates flat image}
\usage{
mod_disc_flat(x, hdrlst, header, mean.method = "mode", sdo.image = "FALSE")
}
\arguments{
\item{x}{tibble containing columns wih pixel coordinates i and j and all
the calculated image information as provided by the disc extraction module.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{mean.method}{string implemented methods are "mean", "median" and
"mode", with "mode" as default.}

\item{sdo.image}{boolean switch for dummy use in the case of non SDO images.}
}
\value{
list containing rescaled flat image, hdrlst and header.
}
\description{
Recalibrates the flat image to \link{0,1} interval. Be aware, that
the resulting intensity contrast values should be multiplied by some
appropriate factor before converting them back to FITS, since in the FITS
image all values are treated as integers.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
