% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fits_import.r
\name{mod_fits_import}
\alias{mod_fits_import}
\title{Imports a single FITS file}
\usage{
mod_fits_import(
  inp_data_path,
  inp_file_name,
  sdo.image = "FALSE",
  cut.image = "FALSE"
)
}
\arguments{
\item{inp_data_path}{string with input data path.}

\item{inp_file_name}{string with input file name.}

\item{sdo.image}{if TRUE the image is originally an imported sdo jpg-file.}

\item{cut.image}{if TRUE the image contains additional text in the corners.}
}
\value{
list with fitsim, hdrlst and header.
}
\description{
Imports single FITS file given the file path and name. Returns
tibble with image, tibble with hdrlst and character vector header.
}
\author{
\href{mailto:thomas.friedli@bluewin.ch}{Thomas K. Friedli}
}
