% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_frame_import.r
\name{mod_frame_import}
\alias{mod_frame_import}
\title{Imports a single image file}
\usage{
mod_frame_import(
  inp_data_path,
  inp_file_name,
  sdo.image,
  parse.method = "SDO/HMI",
  flip.image = "FALSE",
  flop.image = "FALSE"
)
}
\arguments{
\item{inp_data_path}{string with input data path.}

\item{inp_file_name}{string with input file name.}

\item{sdo.image}{boolean if TRUE the image is an imported sdo jpg-file.}

\item{parse.method}{string method for parsing the provided file.name.}

\item{flip.image}{boolean if TRUE the image will be flipped.}

\item{flop.image}{boolean if TRUE the image will be flopped.}
}
\value{
list with fitsim, hdrlst and header.
}
\description{
Imports single image file given the file path and name. Returns
tibble with image, tibble with hdrlst and character vector header. Changes
image header, flips and flops the image, if requested.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
