% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spot_extraction.r
\name{mod_spot_extraction}
\alias{mod_spot_extraction}
\title{Extracts the spot, penumbra and umbra features}
\usage{
mod_spot_extraction(
  x,
  hdrlst,
  header,
  spot.contrast = 0.85,
  umbra.contrast = 0.7
)
}
\arguments{
\item{x}{tibble containing 3 columns wih pixel coordinates i and j and flat
image pixel values.}

\item{hdrlst}{list containing image FITS header keywords and values.}

\item{header}{list containing image FITS header.}

\item{spot.contrast}{cutoff threshold for spots.}

\item{umbra.contrast}{cutoff threshold for umbrae.}
}
\value{
tibble with mask of the extracted features and their contrast values.
}
\description{
Extracts the spot, penumbra and umbra features by applying
cutoff thresholds. Calculates also the hemispheric corrected values for
the extracted feature areas and contrasts.
}
\author{
\href{mailto:thomas.k.friedli@bluewin.ch}{Thomas K. Friedli}
}
