% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_modify_header.r
\name{newKwv}
\alias{newKwv}
\title{Utilities for modifying FITS header entries}
\usage{
newKwv(keyw, val, note = "")
}
\value{
string vector with modified header.
}
\description{
Utilities for modifying FITS header entries.
}
\author{
[Thomas K. Friedli](mailto:thomas.friedli@bluewin.ch)
}
